feat: Implement UPSERT logic for location and rice blast warnings, adjust location ID type, and enhance weather API robustness with IPv4 forcing and error handling.

+++ src/entities/location.entity.ts
@@ -8,7 +8,7 @@ import { ApiProperty } from '@nestjs/swagger';
 @Entity('locations')
 export class Location {
   @ApiProperty({ description: 'ID cố định = 1', example: 1 })
-  @PrimaryColumn({ type: 'bigint', default: 1 })
+  @PrimaryColumn({ type: 'int', default: 1 })
   id!: number;
 
   @ApiProperty({
